feat: App is operational now

+++ bmicalculator/lib/constants.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 const Color kActivatorColor = Color(0xFF1D1A2F);
 const Color kInactivateColor = Color(0xFF121124);
-int kHeight = 180;
+double kHeight = 180;
 
 const kLabelTextStyle = TextStyle(color: Color(0XFFC0BDCB), fontSize: 20);
 const kTextStyle = TextStyle(fontSize: 50, fontWeight: FontWeight.w700);
